pyproject.toml
[project]
name = "streaming-service"
version = "0.1.0"
description = ""
authors = [
    {name = "Boris Averin",email = "89068157313@mail.ru"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "pydantic (>=2.11.4,<3.0.0)",
    "alembic (>=1.15.2,<2.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "redis (>=6.1.0,<7.0.0)",
    "greenlet (>=3.2.2,<4.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
